wip: upgrade to zig-master
the main breakage here is caused by writergate. there's some weirdness
going on to do with file paths and the build system too though, so that
should be dealt with before merging. also the new writer and reader
don't provide any error specificity, so we don't know why a read or a
write failed now. this should also be considered: perhaps creating a
custom interface is warranted?