fix build interface
we set the cwd of dimmer to the script directory such that the run step passes
the correct relative paths. this caused weirdness in zig's dependency
file handling - zig expects relative paths to be relative to the project
root instead of the cwd of the run step. changing paths to be absolute
is a suitable workaround.